requirements.txt
streamlit
pandas
psycopg2-binary
plotly
scikit-learn
tensorflow
kafka-python